pyproject.toml
[project]
name = "commento"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "griotold", email = "gotjd9773@naver.com"},
]
dependencies = [
    "fastapi>=0.121.1",
    "uvicorn>=0.38.0",
    "langchain==0.3.3",
    "langchain-openai==0.2.2",
    "openai==1.51.2",
    "python-dotenv==1.0.1",
    "httpx==0.27.2",
    "transformers>=4.57.1",
    "torch>=2.9.1",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "PyJWT>=2.10.0",
]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false

[tool.pdm.scripts]
start = "uvicorn src.main:app --reload --host 0.0.0.0 --port 8000"
test = "pytest tests/ -v"
test-front = {cmd = "python -m http.server 3000", working_dir = "test-front"}

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "aiosqlite>=0.22.1",
]
